requirements.txt
Python==3.12.7
Django==5.0.7
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0